fix: reaction value on (fresh) page load
